+++ src/wrap.rs
@@ -18,16 +18,70 @@ impl WrapGameInfo {
 pub fn infer_game_from_steam(title_finder: &TitleFinder) -> Option<WrapGameInfo> {
     let app_id = std::env::var("STEAMAPPID").ok()?.parse::<u32>().ok()?;
 
-    log::debug!("Found Steam environment variable: STEAMAPPID={}", app_id,);
+    log::debug!("Found Steam environment variable: STEAMAPPID={}", app_id);
 
     let result = WrapGameInfo {
         name: title_finder.find_one(&[], &Some(app_id), &None, false),
         gog_id: None,
     };
 
-    if result.is_empty() {
-        None
-    } else {
+    (!result.is_empty()).then_some(result)
+}
+
+pub mod lutris {
+    use super::*;
+
+    use crate::path::StrictPath;
+
+    use std::sync::Mutex;
+
+    // TODO: Refactor to avoid shared state.
+    static INFERRED_NORMALIZED_TITLE: Mutex<Option<String>> = Mutex::new(None);
+
+    pub struct Metadata {
+        pub title: String,
+        pub base: Option<StrictPath>,
+        pub prefix: Option<StrictPath>,
+    }
+
+    pub fn infer() -> Option<WrapGameInfo> {
+        let title = std::env::var("game_name").ok()?;
+
+        log::debug!("Found Lutris environment variable: game_name={}", &title);
+
+        let result = WrapGameInfo {
+            name: Some(title),
+            gog_id: None,
+        };
+
         Some(result)
+    }
+
+    pub fn save_normalized_title(title: String) {
+        let Ok(mut guard) = INFERRED_NORMALIZED_TITLE.lock() else {
+            return;
+        };
+        *guard = Some(title.clone());
+    }
+
+    pub fn infer_metadata() -> Option<Metadata> {
+        let guard = INFERRED_NORMALIZED_TITLE.lock().ok()?;
+        let title = (*guard).as_ref()?.clone();
+
+        let base = std::env::var("GAME_DIRECTORY").ok();
+        let prefix = std::env::var("WINEPREFIX").ok();
+
+        log::debug!(
+            "Found Lutris environment variables for inferred game '{}': GAME_DIRECTORY={:?}, WINEPREFIX={:?}",
+            &title,
+            &base,
+            &prefix
+        );
+
+        Some(Metadata {
+            title,
+            base: base.map(StrictPath::new),
+            prefix: prefix.map(StrictPath::new),
+        })
     }
 }